feat: Add refs to useFloatingOptions in ExpandableSelector
This commit adds the `refs` parameter to the `useFloatingOptions` hook in the `ExpandableSelector` component. The `refs` object is now passed as a prop to the `useFloatingOptions` hook and used to focus the reference element when an option is selected.

This modification improves the functionality of the `ExpandableSelector` component by allowing the reference element to receive focus after an option is selected, enhancing the user experience.