refactor: Update PopoverPanel to use FloatingFocusManager for focus trapping
The PopoverPanel component has been refactored to use the FloatingFocusManager from the @floating-ui/react library for focus trapping. This change replaces the previous usage of the FocusTrap component from the focus-trap-react library.

By using the FloatingFocusManager, we ensure that focus is properly managed within the PopoverPanel, allowing for a better user experience when interacting with the panel's content.