refactor: Add 'weeks' and 'months' to the list of properties in organizeSets.ts
This commit adds the 'weeks' and 'months' properties to the list of properties in the organizeSets.ts file. These properties are now included in the array of properties that are being organized for a user.

This change improves the functionality of the organizeSets function by ensuring that the 'weeks' and 'months' properties are properly handled and organized for each user.