feat: Update FocusTitle component to include RhytmicRerender and PageMetaTags
This commit updates the FocusTitle component in the increaser/app/home/components directory. It adds the RhytmicRerender component from the lib/ui/base/RhytmicRerender module and the PageMetaTags component from the lib/next-ui/metadata/PageMetaTags module. The RhytmicRerender component is used to render the PageMetaTags component, which displays the title of the page based on the duration of the focus session.

These changes improve the functionality of the FocusTitle component by dynamically updating the page title based on the duration of the focus session, providing a more informative and engaging user experience.