Focus Page improvements (#20)
* feat: Send email to target users with exciting updates about Increaser and ProductHunt launch

This commit adds a new file, productHunt.ts, which contains the logic to send an email to target users with exciting updates about Increaser. The email includes information about new features and enhancements, as well as a link to the Increaser ProductHunt page. The email is sent to users who have not opted out of receiving emails and who registered before the campaign started.

This feature improves user engagement by keeping them informed about the latest updates and encourages them to support Increaser on ProductHunt.

Co-authored-by: Radzion <radzion@increaser.org>

* feat: Update FocusTitle component to include RhytmicRerender and PageMetaTags

This commit updates the FocusTitle component in the increaser/app/home/components directory. It adds the RhytmicRerender component from the lib/ui/base/RhytmicRerender module and the PageMetaTags component from the lib/next-ui/metadata/PageMetaTags module. The RhytmicRerender component is used to render the PageMetaTags component, which displays the title of the page based on the duration of the focus session.

These changes improve the functionality of the FocusTitle component by dynamically updating the page title based on the duration of the focus session, providing a more informative and engaging user experience.

* feat: Update FocusTitle component to include RhytmicRerender and PageMetaTags

* feat: Update FocusAssistance component to use Menu for notifications settings

This commit updates the FocusAssistance component in the increaser/app/focus/components/FocusAssistance directory. It replaces the LabeledValue component with the Menu component to improve the layout and functionality of the notifications settings. The Menu component provides a more intuitive and user-friendly interface for managing browser and sound notifications.

These changes enhance the user experience by making it easier for users to customize their notification preferences and improve their focus sessions.

* feat: Update FocusSetWidget component and config

This commit updates the FocusSetWidget component in the increaser/app/focus/components/FocusSetWidget directory. It removes the CurrentFocusTask component and updates the import path for the MinimalisticFocusSet component. Additionally, it adds a new config file, config.ts, which exports the focusSetWidgetConfig object with a padding value of 16.

These changes improve the structure and functionality of the FocusSetWidget component, providing a more streamlined and customizable focus set widget.

* feat: Update padding and max-height in Navigation and SidebarContent

This commit updates the padding and max-height properties in the Navigation and SidebarContent components in the increaser/app/ui/Navigation/index.tsx file. The padding value in SidebarContent is modified to remove the bottom padding and the max-height property is added to limit the height of the component. These changes improve the layout and appearance of the navigation and sidebar content, providing a more consistent and visually appealing user interface.

* refactor: Update import paths for Switch component in multiple files

* feat: Update defaultFocusSounds with new options

This commit updates the defaultFocusSounds array in the FocusSound.ts file. It adds new focus sound options, including "Study with me" and "Chill Lo-fi hip hop". These additions provide users with a wider selection of focus sounds to enhance their productivity and concentration.

Co-authored-by: Radzion <radzion@increaser.org>

* feat: Remove YouTube focus sounds and related components

This commit removes the YouTube focus sounds feature and its related components. The persistent state key, YouTubeFocusPreference, is also removed. This change simplifies the codebase and removes unused functionality.

Co-authored-by: Radzion <radzion@increaser.org>

* feat: Update background color in YouTubeFocusMusicPlayer and FocusSetWidget

This commit updates the background color in the YouTubeFocusMusicPlayer and FocusSetWidget components. The background color is changed from "mist" to "foreground" in YouTubeFocusMusicPlayer and from "transparent" to "foreground" in FocusSetWidget. These changes improve the visual consistency and enhance the user interface of the application.

Co-authored-by: Radzion <radzion@increaser.org>

* feat: Update grid-template-columns in SoundItem component

This commit updates the grid-template-columns property in the SoundItem component of the YouTubeFocusList. The value is changed from "32px 1fr auto" to "minmax(24px, auto) 1fr auto". This change improves the layout and alignment of the component, providing a more consistent and visually appealing user interface.

* feat: Remove SoundsPlayer component

The SoundsPlayer component in the `youTube/FocusSoundsList` directory has been removed. This change simplifies the codebase and removes unused functionality.

* refactor: Update useStateCorrector to improve state correction logic

The `useStateCorrector` function in the `lib/ui/state/useStateCorrector.ts` file has been updated to improve the state correction logic. This change ensures that the correct state is returned and set based on the provided correction function. The `useStateCorrector` function is used in the `useYouTubeFocusPreference` hook in the `increaser/app/focus/audio/youTube/state/useYouTubeFocusPreference.ts` file to correct the state based on the `focusSounds` array. This refactor enhances the functionality and maintainability of the codebase.

Co-authored-by: Radzion <radzion@increaser.org>

* feat: Update FocusAssistance component to use Menu for notifications settings

---------

Co-authored-by: Radzion <radzion@increaser.org>