Feat/focus 4 (#21)
* feat: Add focus page

* feat: Update FocusTitle component to include RhytmicRerender and PageMetaTags

* feat: Improve expand/collapse functionality in HabitsAdvice component

* feat: Add lazy loading for images in NavigateToPlanDecoration and update PlanPage layout

The code changes in this commit add lazy loading for images in the `NavigateToPlanDecoration` component, improving page load performance. Additionally, the layout of the `PlanPage` component is updated to include the `HabitsReview` and `TasksReview` components, providing a more comprehensive view of the user's habits and tasks.

* feat: Add lazy loading for images in NavigateToPlanDecoration and update PlanPage layout

* feat: Add isOverdue property to TaskDescription type and update getDemoTasks function

This commit adds the `isOverdue` property to the `TaskDescription` type in the `tasks.ts` file. The `isOverdue` property is used to indicate whether a task is overdue or not. Additionally, the `getDemoTasks` function in the same file is updated to include the `isOverdue` property when creating task objects.

These changes enhance the functionality of the `TaskDescription` type and the `getDemoTasks` function, allowing for better tracking and management of task deadlines.

* feat: Update WaterfallIcon viewBox and paths

* feat: Add lazy loading for images in NavigateToPlanDecoration and update PlanPage layout

* feat: Add lazy loading for images in NavigateToPlanDecoration and update PlanPage layout