Feat/improvements (#25)
* Refactor page components to remove makeWebsitePage function

* feat: Add lazy loading for images in Sidebar NavigationToInternalPage component

* feat: Add membership page and components

This commit adds the membership page and related components to the application. The membership page allows users to access exclusive features and benefits by subscribing to a membership plan. The components include NoMembershipDecoration, MembershipPage, MembershipOverview, MembershipNavigationDecoration, FreeTrialDecoration, and related styling and logic.

Co-authored-by: Radzion <radzion@increaser.org>

* feat: Add lazy loading for images in CrownIcon and InstallPrompt components

* feat: Refactor file structure and update component imports

This commit refactors the file structure by moving the SubscriptionBillingCycleSelector component from the increaser/app/membership/components directory to the increaser/ui/subscription directory. It also updates the component imports in the MembershipOffer and MembershipPage components to reflect the new file location. This change improves the organization and maintainability of the codebase.

Co-authored-by: Radzion <radzion@increaser.org>

---------

Co-authored-by: Radzion <radzion@increaser.org>