Feat/goals (#26)
* feat: Add lazy loading for images in Sidebar NavigationToInternalPage component

* feat: Add lazy loading for images in Sidebar NavigationToInternalPage component

* feat: Add lazy loading for images in Sidebar NavigationToInternalPage component

* feat: Add lazy loading for images in Sidebar NavigationToInternalPage component

* feat: Add 'coldShower' habit to habit tracker

This commit adds the 'coldShower' habit to the habit tracker in the 'habits.ts' file. The 'coldShower' habit is a daily practice that involves taking a cold shower to boost energy levels, improve circulation, enhance mental clarity, reduce stress, and improve mood. This addition expands the range of habits that users can track and encourages a healthy lifestyle.

Co-authored-by: Radzion <radzion@increaser.org>

* feat: Update tsconfig.json to remove unnecessary file inclusion

This commit updates the tsconfig.json file to remove the unnecessary inclusion of the "../../lib/dnd/DnDList.tsx" file. This file is no longer needed in the project and can be safely excluded. By removing this unnecessary inclusion, the project's build process will be more efficient and faster.

Co-authored-by: Radzion <radzion@increaser.org>

---------

Co-authored-by: Radzion <radzion@increaser.org>