Feat/improvements 2 (#27)
* feat: Add lazy loading for images in Sidebar NavigationToInternalPage component

* feat: Add Education entity and migration to set completedEducation to empty array for all users

This commit adds the Education entity to the increaser/entities directory. The Education entity defines the possible values for the completedEducation property in the User entity. It includes two options: 'vision' and 'goals'.

The commit also adds a migration file, setEducation.ts, to the increaser/db/migrations directory. This migration sets the completedEducation property to an empty array for all users in the database.

These changes enable the tracking of completed education milestones for users and ensure that all users start with an empty array for completed education.

Co-authored-by: Radzion <radzion@increaser.org>

* feat: Add lazy loading for images in ProductEducationBlock component

* feat: Add lazy loading for images in ProductEducationBlock component

* feat: Add lazy loading for images in ExpandableSectionListTitle and GoalItem components

* feat: Add goals page and components

This commit adds the goals page and related components to the application. The goals page allows users to define and track their goals. The components include GoalsPage, GoalItem, VisionAttributeItem, and related styling and logic.

Co-authored-by: Radzion <radzion@increaser.org>

* feat: Set completedEducation to empty array for all users

This commit adds a migration file, setEducation.ts, which sets the completedEducation property to an empty array for all users in the database. This change enables the tracking of completed education milestones for users and ensures that all users start with an empty array for completed education.

Co-authored-by: Radzion <radzion@increaser.org>

* feat: Add lazy loading for images in VisionAttributeItem component

* feat: Add deadlineAt property to Goal entity and related components

This commit adds the deadlineAt property to the Goal entity in the increaser/entities/Goal.ts file. The deadlineAt property represents the deadline for achieving the goal and is optional.

The commit also updates the GoalFormShape and GoalForm components in the increaser/ui/goals/form directory to include the deadlineAt property.

These changes enable users to set a deadline for their goals and track their progress more effectively.

Co-authored-by: Radzion <radzion@increaser.org>

* feat: Update MembershipPersuasion component with dynamic product name

This commit updates the MembershipPersuasion component in the increaser/app/membership/components/MembershipPersuasion.tsx file. It adds a dynamic product name by importing the `productName` variable from the `@increaser/config` module. The product name is then used in the SectionTitle component to display a personalized message to the user.

This change enhances the user experience by making the product name more dynamic and customizable.

Co-authored-by: Radzion <radzion@increaser.org>

* feat: Add GridIcon component and update related files

This commit adds the GridIcon component to the lib/ui/icons directory. The GridIcon component is an SVG icon that represents a grid layout. It is used in various parts of the application to visually represent a grid or grid-related functionality.

The commit also updates the following files to incorporate the GridIcon component:
- lib/ui/icons/GridIcon.tsx
- increaser/app/vision/VisionPage.tsx
- increaser/website/landing/demo/ProductToolSlice.tsx
- increaser/ui/subscription/SubscriptionFeatures.tsx
- increaser/app/ui/Navigation/navigationPathInfo.tsx
- increaser/ui/tools/productToolIconRecord.tsx
- increaser/website/landing/demo/VisionSlice.tsx
- increaser/entities/ProductTool.ts

These changes enhance the visual consistency and usability of the application by introducing a new icon component and updating the relevant files to use it.

Co-authored-by: Radzion <radzion@increaser.org>

* feat: Add lazy loading for images in various components

* feat: Add lazy loading for images in VisionAttributeItemContent component

* feat: Update AWS SDK dependencies and add lazy loading for images

This commit updates the AWS SDK dependencies in the package.json file to the latest stable versions. Specifically, it updates "@aws-sdk/client-s3" to "^3.590.0" and adds a new dependency "@aws-sdk/s3-request-presigner" with version "^3.590.0".

Additionally, this commit introduces lazy loading for images in various components:
- VisionAttributeItemContent component
- VisionAttributeItem component
- ExpandableSectionListTitle component
- GoalItem component
- ProductEducationBlock component

These changes improve the performance of the application by loading images only when necessary, reducing the initial page load time.

Co-authored-by: Radzion <radzion@increaser.org>

* feat: Update AWS SDK dependencies and add lazy loading for images

* feat: Add lazy loading for images in VisionImageInput component

* feat: Update AWS SDK dependencies and add lazy loading for images

This commit updates the AWS SDK dependencies in the package.json file to the latest stable versions. Specifically, it updates "@aws-sdk/client-s3" to "^3.590.0" and adds a new dependency "@aws-sdk/s3-request-presigner" with version "^3.590.0".

Additionally, this commit introduces lazy loading for images in various components.

* feat: Update AWS SDK dependencies and add lazy loading for images

This commit updates the AWS SDK dependencies and introduces lazy loading for images in various components. It also adds a new dependency "@aws-sdk/s3-request-presigner" with version "^3.590.0". These changes improve the performance of the application by loading images only when necessary, reducing the initial page load time.

* feat: Update Vision entity to remove imageUrl and add imageId property

This commit modifies the Vision entity by removing the imageUrl property and adding the imageId property. The imageUrl property was no longer needed, and the imageId property provides a more efficient way to reference images. This change improves the structure and organization of the Vision entity.

Co-authored-by: Radzion <radzion@increaser.org>

* refactor: Update deletePublicBucketFile to use getPublicBucketName

This commit refactors the deletePublicBucketFile function in the deletePublicBucketFile.ts file. It replaces the usage of the getEnvVar function to retrieve the PUBLIC_BUCKET_NAME environment variable with the getPublicBucketName function. This change improves the readability and maintainability of the code by using a more descriptive function name.

Co-authored-by: Radzion <radzion@increaser.org>

* feat: Introduce VisionImageInputContent component for displaying image upload status

---------

Co-authored-by: Radzion <radzion@increaser.org>