Feat/goal emoji (#32)
* feat: Refactor CreateTaskForm and EditTaskForm components

The CreateTaskForm and EditTaskForm components have been refactored to remove the unnecessary VStack component and simplify the structure. This improves the readability and maintainability of the code.

* Refactor CreateTaskForm and EditTaskForm components

* feat: Add emoji support for goals in EditGoalForm

The EditGoalForm component has been updated to include support for setting an emoji for a goal. This allows users to add a visual representation to their goals, improving the user experience and personalization of the application.

* Refactor ExpandableInputOpener component to use textInputHeight instead of defaultInputHeight